pyproject.toml
[tool.poetry]
name = "royston"
version = "0.0.13"
description = "A real time trend detection algorithm"
authors = ["Ian Read <ianharveyread@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
test = 'scripts:test'
coverage = 'scripts:coverage'
focused = 'scripts:test_focused'

[tool.poetry.dependencies]
dateparser = "^1.0.0"
gensim = "^4.0.1"
nltk = "^3.6.2"
pytest = "^6.2.5"
pytest-cov = "^2.12.1"
python = "^3.8"
python-dateutil = "^2.8.1"
pytz = "^2021.1"
coverage = "^5.5"
twine = "^3.4.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
black = "^21.9b0"

[tool.black]
line-length = 79

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
